requirements.txt
Flask
sentence-transformers
pdfplumber
python-docx
PyMuPDF
